Removed out-of-sync lockfile check in pr.yml
Since we're using matrix test strategy lockfile will be different.
This also means that the changed files likely don't pass lint/fmt/check,
so we'll remove those as well.
We'll have to rely on developers using the `rye run all` command in pyproject.toml.
Could also set up pre-commit, but since this isn't multi-contributor we'll forgo that for now.